lealang_chalk_rs
Struct Chalk

pub struct Chalk { /* private fields */ }

Implementations

impl Chalk

For all Chalks with different color types

pub fn new() -> Self

Creates a Chalk with a black background and a white foreground

Example

use chalk_rs::Chalk;

let mut chalk = Chalk::new();
// the chalk can be used here

pub fn string(&self, string: &dyn ToString) -> String

Formats a string using the style of the given Chalk.

This will return the string after being formatted to the console. When using string literals, please use a reference.

For example:

chalk.yellow().string(&"this is yellow");

Arguments

string - The item to print out. It must implement ToString and should be a reference

Example

use chalk_rs::Chalk;

let mut chalk = Chalk::new();
let text = chalk.yellow().string(&"this is yellow");

pub fn print(&self, string: &dyn ToString) -> String

Prints a string using the style of the given chalk.

This will return the text that was outputted to the console. When using string literals, please use a reference.

For example:

chalk.yellow().print(&"this is yellow");

Arguments

string - The item to format and output. It must implement ToString and should be a reference.

Example

use chalk_rs::Chalk;

let mut chalk = Chalk::new();
chalk.yellow().print(&"this is yellow");

pub fn println(&self, string: &dyn ToString) -> String

Prints a string using the style of the given chalk on a new line.

This will return the text that was outputted to the console. When using string literals, please use a reference.

For example:

chalk.yellow().println(&"this is yellow");

Arguments

string - The item to format and output. It must implement ToString and should be a reference.

Example

use chalk_rs::Chalk;

let mut chalk = Chalk::new();
chalk.yellow().println(&"this is yellow");

impl Chalk

pub fn default_color(&mut self) -> &mut Self

pub fn black(&mut self) -> &mut Self

pub fn red(&mut self) -> &mut Self

pub fn green(&mut self) -> &mut Self

pub fn yellow(&mut self) -> &mut Self

pub fn blue(&mut self) -> &mut Self

pub fn magenta(&mut self) -> &mut Self

pub fn cyan(&mut self) -> &mut Self

pub fn light_gray(&mut self) -> &mut Self

pub fn light_grey(&mut self) -> &mut Self

pub fn gray(&mut self) -> &mut Self

pub fn grey(&mut self) -> &mut Self

pub fn light_black(&mut self) -> &mut Self

pub fn light_red(&mut self) -> &mut Self

pub fn light_green(&mut self) -> &mut Self

pub fn light_yellow(&mut self) -> &mut Self

pub fn light_blue(&mut self) -> &mut Self

pub fn light_magenta(&mut self) -> &mut Self

pub fn light_cyan(&mut self) -> &mut Self

pub fn white(&mut self) -> &mut Self

pub fn ansi(&mut self, color: u8) -> &mut Self

pub fn rgb(&mut self, r: u8, g: u8, b: u8) -> &mut Self

impl Chalk

pub fn default_background(&mut self) -> &mut Self

pub fn bg_black(&mut self) -> &mut Self

pub fn bg_red(&mut self) -> &mut Self

pub fn bg_green(&mut self) -> &mut Self

pub fn bg_yellow(&mut self) -> &mut Self

pub fn bg_blue(&mut self) -> &mut Self

pub fn bg_magenta(&mut self) -> &mut Self

pub fn bg_cyan(&mut self) -> &mut Self

pub fn bg_light_gray(&mut self) -> &mut Self

pub fn bg_light_grey(&mut self) -> &mut Self

pub fn bg_gray(&mut self) -> &mut Self

pub fn bg_grey(&mut self) -> &mut Self

pub fn bg_light_black(&mut self) -> &mut Self

pub fn bg_light_red(&mut self) -> &mut Self

pub fn bg_light_green(&mut self) -> &mut Self

pub fn bg_light_yellow(&mut self) -> &mut Self

pub fn bg_light_blue(&mut self) -> &mut Self

pub fn bg_light_magenta(&mut self) -> &mut Self

pub fn bg_light_cyan(&mut self) -> &mut Self

pub fn bg_white(&mut self) -> &mut Self

pub fn bg_ansi(&mut self, color: u8) -> &mut Self

pub fn bg_rgb(&mut self, r: u8, g: u8, b: u8) -> &mut Self

impl Chalk

pub fn reset_style(&mut self) -> &mut Self

pub fn reset_weight(&mut self) -> &mut Self

pub fn bold(&mut self) -> &mut Self

pub fn dim(&mut self) -> &mut Self

pub fn italic(&mut self) -> &mut Self

pub fn unitalic(&mut self) -> &mut Self

pub fn no_underline(&mut self) -> &mut Self

pub fn underline(&mut self) -> &mut Self

pub fn double_underline(&mut self) -> &mut Self

pub fn stop_blink(&mut self) -> &mut Self

pub fn blink(&mut self) -> &mut Self

pub fn invert(&mut self) -> &mut Self

pub fn uninvert(&mut self) -> &mut Self

pub fn hide(&mut self) -> &mut Self

pub fn unhide(&mut self) -> &mut Self

pub fn is_normal_weight(&self) -> bool

pub fn is_bold(&self) -> bool

pub fn is_dim(&self) -> bool

pub fn is_italicized(&self) -> bool

pub fn has_underlines(&self) -> bool

pub fn is_single_underlined(&self) -> bool

pub fn is_double_underlined(&self) -> bool

pub fn is_blinking(&self) -> bool

pub fn is_inverted(&self) -> bool

pub fn is_hidden(&self) -> bool

pub fn num_underlines(&self) -> u8